Form cubits folder reorganization
