+++ lib/logic/caregiver/forms/badge/badge_form_state.dart
