+++ lib/pages/caregiver/forms/badge_form_page.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
-import 'package:fokus/logic/caregiver/badge_form/badge_form_cubit.dart';
+import 'package:fokus/logic/caregiver/forms/badge/badge_form_cubit.dart';
 import 'package:fokus/model/ui/form/badge_form_model.dart';
 import 'package:fokus/utils/ui/dialog_utils.dart';
 import 'package:fokus/utils/ui/form_config.dart';
